Compute the separators boolean
This is set in function of elem_size. Basically, if there is
enough room to print an element + 2 spaces + 2 vertical lines
(aka separators), then separators are printed. It is also taken
into account in the 'print_elem' function since it changes the
coordinates at which an element must be printed.